+++ index.js
@@ -22,3 +22,4 @@ module.exports.errors = process.requireApi('lib/errors/index.js');
 module.exports.socket = process.requireApi('lib/socket/index.js');
 module.exports.emitters = process.requireApi('lib/emitters/index.js');
 module.exports.grunt = process.requireApi('lib/grunt/index.js');
+module.exports.multipart = process.requireApi('lib/multipart/index.js');