+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 4.2.0 /
+
+## NEW FEATURES
+
+- Add require('@openveo/api').util.areSameArrays to shallow validates that two arrays contain the same values, no more no less
+
 # 4.1.0 / 2017-09-12
 
 ## DEPENDENCIES